Use clone instead of merge for docs

+++ scripts/push-documentation.sh
@@ -2,26 +2,19 @@
 set -e # Exit with nonzero exit code if anything fails
 
 # Pull requests and commits to other branches shouldn't try to deploy
-# if [ "$TRAVIS_PULL_REQUEST" != "false" -o "$TRAVIS_BRANCH" != "master" ]; then
-#     echo "Skipping documentation deploy."
-#     exit 0
-# fi
-
-# git checkout docs
-# git merge master -X theirs --no-ff --no-commit
-# git reset 
-# git add .
-# git commit -m "Update docs branch" 
-# git push origin docs
+if [ "$TRAVIS_PULL_REQUEST" != "false" -o "$TRAVIS_BRANCH" != "master" ]; then
+    echo "Skipping documentation deploy."
+    exit 0
+fi
 
 rm -rf docs
 git clone -b docs git@github.com:gini/gini-vision-lib-ios.git docs
 
 rm -rf docs/*
-cp -a Documentation/. docs/
+cp -a Documentation/. docs/Documentation/
 cd docs
 
 git add -u
 git add .
 git diff --quiet --exit-code --cached || git commit -a -m 'Deploy Gini Vision Library for iOS documentation to docs branch'
-git push
+git push origin docs
